Add actions manager and base planning logic

+++ Assets/Scripts/Components/ResourceMiningPointComponent.cs
@@ -4,6 +4,6 @@ using Entitas;
 public sealed class ResourceMiningComponent : IComponent
 {
     public float ActionIntervalDelay;
-    public int ResourceCountPerInterval;
+    public ushort ResourceCountPerInterval;
     public ResourceType RequirementsForInteraction;
 }
